+++ CHANGES.md
@@ -14,6 +14,7 @@ Core Grammars:
 - enh(swift) support `macro` keyword [Bradley Mackey][]
 - enh(swift) support parameter pack keywords [Bradley Mackey][]
 - enh(swift) regex literal support [Bradley Mackey][]
+- enh(swift) `@unchecked` and `@Sendable` support [Bradley Mackey][]
 
 Dev tool: